Fix git pull problem on AWS Smoketest. Change engines --PQ value. Now it is calculated from number of cores. Add build.summary file for skipped PR-s to enable list them in listTests.sh. Fix showStatus.py problem if multiple Regression result file found.
Deploy on 2021-04-30 22:15:36

+++ smoketest.sh
@@ -203,8 +203,11 @@ echo "---------------------------------"
 
 echo "Find and remove 'build' and ' HPCC-Platfrom' directories"
 
-find PR-*/ -maxdepth 1 -iname 'build' -type d -print -exec rm -rf '{}' \;
-find PR-*/ -maxdepth 1 -iname 'HPCC-Platform' -type d -print -exec rm -rf '{}' \;
+if [[ "$keepFiles" != "1" ]]
+then
+    find PR-*/ -maxdepth 1 -iname 'build' -type d -print -exec rm -rf '{}' \;
+    find PR-*/ -maxdepth 1 -iname 'HPCC-Platform' -type d -print -exec rm -rf '{}' \;
+fi
 
 find OldPrs/PR-*/ -maxdepth 1 -iname 'build' -type d -print -exec rm -rf '{}' \;
 find OldPrs/PR-*/ -maxdepth 1 -iname 'HPCC-Platform' -type d -print -exec rm -rf '{}' \;
